requirements.txt
fastapi
uvicorn[standard]
python-multipart
aiofiles
assemblyai>=0.33.0
httpx>=0.27.0
httpcore>=1.0.0
supabase
python-dotenv
groq
itsdangerous
werkzeug
requests
jinja2
pandas
openpyxl
